+++ Spartane.Web/Areas/WebApiConsumer/Tipo_de_Equipo_Tactico/ITipo_de_Equipo_TacticoApiConsumer.cs
@@ -14,19 +14,19 @@ namespace Spartane.Web.Areas.WebApiConsumer.Tipo_de_Equipo_Tactico
         Int32 SelCount();
         ApiResponse<IList<Spartane.Core.Domain.Tipo_de_Equipo_Tactico.Tipo_de_Equipo_Tactico>> SelAll(Boolean ConRelaciones);
         ApiResponse<IList<Spartane.Core.Domain.Tipo_de_Equipo_Tactico.Tipo_de_Equipo_Tactico>> SelAllComplete(Boolean ConRelaciones);
-        ApiResponse<Spartane.Core.Domain.Tipo_de_Equipo_Tactico.Tipo_de_Equipo_Tactico> GetByKey(short Key, Boolean ConRelaciones);
-        ApiResponse<Spartane.Core.Domain.Tipo_de_Equipo_Tactico.Tipo_de_Equipo_TacticoPagingModel> GetByKeyComplete(short Key);
-        ApiResponse<bool> Delete(short Key, Spartane.Core.Domain.User.GlobalData Tipo_de_Equipo_TacticoInformation, DataLayerFieldsBitacora DataReference);
-        ApiResponse<Int16> Insert(Spartane.Core.Domain.Tipo_de_Equipo_Tactico.Tipo_de_Equipo_Tactico entity, Spartane.Core.Domain.User.GlobalData Tipo_de_Equipo_TacticoInformation, DataLayerFieldsBitacora DataReference);
-        ApiResponse<Int16> Update(Spartane.Core.Domain.Tipo_de_Equipo_Tactico.Tipo_de_Equipo_Tactico entity, Spartane.Core.Domain.User.GlobalData Tipo_de_Equipo_TacticoInformation, DataLayerFieldsBitacora DataReference);
+        ApiResponse<Spartane.Core.Domain.Tipo_de_Equipo_Tactico.Tipo_de_Equipo_Tactico> GetByKey(int Key, Boolean ConRelaciones);
+        ApiResponse<Spartane.Core.Domain.Tipo_de_Equipo_Tactico.Tipo_de_Equipo_TacticoPagingModel> GetByKeyComplete(int Key);
+        ApiResponse<bool> Delete(int Key, Spartane.Core.Domain.User.GlobalData Tipo_de_Equipo_TacticoInformation, DataLayerFieldsBitacora DataReference);
+        ApiResponse<Int32> Insert(Spartane.Core.Domain.Tipo_de_Equipo_Tactico.Tipo_de_Equipo_Tactico entity, Spartane.Core.Domain.User.GlobalData Tipo_de_Equipo_TacticoInformation, DataLayerFieldsBitacora DataReference);
+        ApiResponse<Int32> Update(Spartane.Core.Domain.Tipo_de_Equipo_Tactico.Tipo_de_Equipo_Tactico entity, Spartane.Core.Domain.User.GlobalData Tipo_de_Equipo_TacticoInformation, DataLayerFieldsBitacora DataReference);
 
         ApiResponse<IList<Spartane.Core.Domain.Tipo_de_Equipo_Tactico.Tipo_de_Equipo_Tactico>> SelAll(Boolean ConRelaciones, Int32 CurrentRecordInt32, Int32 RecordsDisplayedInt32);
         ApiResponse<IList<Spartane.Core.Domain.Tipo_de_Equipo_Tactico.Tipo_de_Equipo_Tactico>> SelAll(Boolean ConRelaciones, string Where, string Order);
         ApiResponse<IList<Spartane.Core.Domain.Tipo_de_Equipo_Tactico.Tipo_de_Equipo_Tactico>> ListaSelAll(Boolean ConRelaciones, string Where, string Order);
         ApiResponse<Spartane.Core.Domain.Tipo_de_Equipo_Tactico.Tipo_de_Equipo_TacticoPagingModel> ListaSelAll(int startRowIndex, int maximumRows, string Where, string Order);
         ApiResponse<IList<Spartane.Core.Domain.Tipo_de_Equipo_Tactico.Tipo_de_Equipo_Tactico>> ListaSelAll(Boolean ConRelaciones, string Where);
-		ApiResponse<short> GenerateID();
-		ApiResponse<short> Update_Datos_Generales(Spartane.Core.Domain.Tipo_de_Equipo_Tactico.Tipo_de_Equipo_Tactico_Datos_Generales entity);
+		ApiResponse<int> GenerateID();
+		ApiResponse<int> Update_Datos_Generales(Spartane.Core.Domain.Tipo_de_Equipo_Tactico.Tipo_de_Equipo_Tactico_Datos_Generales entity);
 		ApiResponse<Spartane.Core.Domain.Tipo_de_Equipo_Tactico.Tipo_de_Equipo_Tactico_Datos_Generales> Get_Datos_Generales(string Key);
